fortify: Explicity disable Clang support Clang has never correctly compiled the FORTIFY_SOURCE defenses due to a couple bugs: Eliding inlines with matching __builtin_* names https://bugs.llvm.org/show_bug.cgi?id=50322 Incorrect __builtin_constant_p() of some globals https://bugs.llvm.org/show_bug.cgi?id=41459
In the process of making improvement to the FORTIFY_SOURCE defenses, the first (silent) bug (coincidentally) becomes worked around, bug exposes the latter which breaks the build. As such, Clang must not be used with CONFIG_FORTIFY_SOURCE until at least latter bug is fixed (in clang 13), and fortify rountines have been rearranged.
Update the kconfig to reflect the reality of the current situation.